starter_ai_agents/camel_ai_starter/pyproject.toml
[project]
name = "camel-ai-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "camel-ai==0.2.75a6",
    "python-dotenv",
    "matplotlib",
]
